pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "uvicorn[standard]>=0.35.0",
]
